devicetree: Enable generation of __symbols__ in all dtb files
In order for a dtb file to be useful with all types of overlays, it
needs to be generated with the -@ flag passed to dtc so that __symbols__
are generated.  This however is not free, and increases the resulting
dtb file by up to approximately 50% today.  In the current worst case
this is moving from 88KiB to 133KiB.  In talking with Frank about this,
he outlined 3 possible ways (with the 4th option of something else
entirely).

1. Make passing -@ to dtc be dependent upon some CONFIG symbol.
2. In the kernel, if the kernel does not have overlay support, discard
the __symbols__ information that we've been passed.
3. Have the bootloader pass in, or not, __symbols__ information.

This patch is an attempt to implement something between the 3rd option
and a different, 4th option.  Frank was thinking that we might introduce
a new symbol to control generation of __symbol__ information for option
1.  I think this gets the usage backwards and will lead to confusion
among users and developers.

My proposal is that we do not want __symbols__ existence to be dependent
on some part of the kernel configuration for a number of reasons.
First, this is out of step with the rest of how dtbs are created today
and more importantly, thought about.  Today, all dtb content is
independent of CONFIG options.  If you build a dtb from a given kernel
tree, everyone will agree on the result.  This is part of the "contract"
on passing old kernels and new dtb files even.

Second, I think this is out of step with how a lot of overlay usage will
occur.  My thinking is that with maximally useful overlays being
available in mainline, lots of use-cases that we have today that result
in a number of DTS files being included can become just overlays.  This
is true in terms of not only evaluation kits but also when these systems
are turned into custom hardware.  This is even more true for SoM based
systems where a physical widget would be a SoM + carrier overlay +
custom parts overlay.  These cases are going to be resolved with
overlays being applied outside of the kernel.

Signed-off-by: Tom Rini <trini@konsulko.com>